CSHARP-6017: Merge LeftJoin translation into JoinMethodToPipelineTranslator
