+++ src/MongoDB.Driver/Linq/Linq3Implementation/Translators/ExpressionToPipelineTranslators/ExpressionToPipelineTranslator.cs
@@ -51,9 +51,8 @@ namespace MongoDB.Driver.Linq.Linq3Implementation.Translators.ExpressionToPipeli
                 case "GroupJoin":
                     return GroupJoinMethodToPipelineTranslator.Translate(context, methodCallExpression);
                 case "Join":
-                    return JoinMethodToPipelineTranslator.Translate(context, methodCallExpression);
                 case "LeftJoin":
-                    return LeftJoinMethodToPipelineTranslator.Translate(context, methodCallExpression);
+                    return JoinMethodToPipelineTranslator.Translate(context, methodCallExpression);
                 case "Lookup":
                     return LookupMethodToPipelineTranslator.Translate(context, methodCallExpression);
                 case "OfType":
